+++ client/src/features/event/EventPage.tsx
@@ -45,6 +45,7 @@ const EventPage: React.FC<{ data: IEvent }> = ({ data }) => {
           </ErrorBoundary>
           <ErrorBoundary FallbackComponent={ErrorFallback}>
             <TimelineContainer
+              key={data.id}
               searchTags={timelines}
               providedHalLinks={event.json._links}
             />