+++ client/src/lib/parse/data/helper.ts
@@ -28,7 +28,6 @@ export const capitalizeLabels = (text: string): string => {
     return text
   }
   return text
-    .toLowerCase()
     .split(' ')
     .map((s) => s.charAt(0).toUpperCase() + s.substring(1))
     .join(' ')