+++ client/src/features/personAndGroup/About.tsx
@@ -118,7 +118,7 @@ const About: React.FC<IProps> = ({ data }) => {
             )),
           )}
         <LinkContainer
-          label="Member Of"
+          label="Member Of (Past or Present)"
           content={memberOf}
           expandColumns
           itemSpacing="single"