+++ client/CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- Fixed the selection of class types in advanced search ([#712](https://github.com/project-lux/lux-frontend/issues/712)).
+
 ## v1.53.1 - 2025-6-23
 
 ### Fixed