+++ client/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added subject_of links to Concept pages About section ([#676](https://github.com/project-lux/lux-frontend/issues/676)).
+
 ### Changed
 
 ### Removed