+++ client/CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Changed the capitalization function for labels ([#700](https://github.com/project-lux/lux-frontend/issues/700)).
+- Changed the label "Member Of" to "Member Of (Past or Present)" ([#701](https://github.com/project-lux/lux-frontend/issues/701)).
 
 ### Removed